👌 Rename logger package to logging
